backend/requirements.txt
# ─── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# ─── Auth ─────────────────────────────────────────────────────────────────────
bcrypt>=4.1.0
PyJWT>=2.8.0

# ─── Validation ───────────────────────────────────────────────────────────────
pydantic[email]>=2.0.0

# ─── File Parsing ─────────────────────────────────────────────────────────────
PyPDF2>=3.0.0
python-docx>=1.1.0

# ─── Environment ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── AI Providers ─────────────────────────────────────────────────────────────
google-generativeai>=0.7.0
groq>=0.9.0

# ─── ML / Embeddings ──────────────────────────────────────────────────────────
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# ─── Utilities ────────────────────────────────────────────────────────────────
requests>=2.31.0

# ─── Audio Analysis ─────────────────────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.0
openai-whisper>=20231117  # Optional: for speech-to-text transcription
